Move animation machine to gameMachine and add update event
