+++ src/Game/animations/index.ts
@@ -1,2 +1,2 @@
-export { makeShowNumber } from './showNumber';
+export { makeShowNumber } from './showNumberMachine';
 export * from './animationMachine';
